See Description
Adding script to switch StarGrazer to the boss paths
Adding in the "Ghost" vertical section to the level 4 whitebox. Couldn't get the ghosts to destroy themselves properly for some reason, so I just had them go way up on their path out of frame. Need to implement "ghosts actually damage the player" still.

There's a weird bug on level 3 that I can't manage to figure out, but it might have something to do with the UI. It's trying to invole EnemyMainPath but failing, and giving the error "EndLayoutGroup: BeginLayoutGroup must be called first"